4: user authentication with express-session + thanks for florian

+++ express-js/4.Express-Typescript/src/controllers/tasksController.ts
@@ -1,15 +1,20 @@
 import { RequestHandler } from "express";
 import mongoose from "mongoose";
 import createHttpError from "http-errors";
-import Task from "../models/tasks";
+import Task from "../models/task";
 import { Params, TaskBody } from "../types/types";
+import { assertIsDefined } from "../assertIsDefined";
 
 /**
  * GET - All Tasks
  */
 export const getTasks: RequestHandler = async (req, res, next) => {
+    const authenticateUserId = req.session.userId;
+
     try {
-        const tasks = await Task.find().exec();
+        assertIsDefined(authenticateUserId);
+
+        const tasks = await Task.find({ userId: authenticateUserId }).exec();
 
         res.status(200).json(tasks);
     } catch (error) {
@@ -22,8 +27,11 @@ export const getTasks: RequestHandler = async (req, res, next) => {
  */
 export const getTask: RequestHandler = async (req, res, next) => {
     const { taskId } = req.params;
+    const authenticateUserId = req.session.userId;
 
     try {
+        assertIsDefined(authenticateUserId);
+
         if (!mongoose.isValidObjectId(taskId)) {
             throw createHttpError(400, "Invalid task id");
         }
@@ -31,7 +39,11 @@ export const getTask: RequestHandler = async (req, res, next) => {
         const task = await Task.findById(taskId).exec();
 
         if (!task) {
-            throw createHttpError(404, "Task not found!"); 
+            throw createHttpError(404, "Task not found!");
+        }
+
+        if (!task.userId.equals(authenticateUserId)) {
+            throw createHttpError(401, "You cannot access to this task!");
         }
 
         res.status(200).json(task);
@@ -43,16 +55,25 @@ export const getTask: RequestHandler = async (req, res, next) => {
 /**
  * POST
  */
-export const postTask: RequestHandler<unknown, unknown, TaskBody, unknown> = async (req, res, next) => {
+export const postTask: RequestHandler<
+    unknown,
+    unknown,
+    TaskBody,
+    unknown
+> = async (req, res, next) => {
     const title = req.body.title;
     const description = req.body.description;
+    const authenticateUserId = req.session.userId;
 
     try {
+        assertIsDefined(authenticateUserId);
+
         if (!title) {
             throw createHttpError(400, "Title cannot be empty!");
         }
 
         const newTask = await Task.create({
+            userId: authenticateUserId,
             title,
             description,
         });
@@ -66,12 +87,20 @@ export const postTask: RequestHandler<unknown, unknown, TaskBody, unknown> = asy
 /**
  * PATCH
  */
-export const patchTask: RequestHandler<Params, unknown, TaskBody, unknown> = async (req, res, next) => {
+export const patchTask: RequestHandler<
+    Params,
+    unknown,
+    TaskBody,
+    unknown
+> = async (req, res, next) => {
     const { taskId } = req.params;
     const updatedTitle = req.body.title;
     const updatedDescription = req.body.description;
-    
+    const authenticateUserId = req.session.userId;
+
     try {
+        assertIsDefined(authenticateUserId);
+
         if (!mongoose.isValidObjectId(taskId)) {
             throw createHttpError(400, "Invalid task Id");
         }
@@ -86,12 +115,15 @@ export const patchTask: RequestHandler<Params, unknown, TaskBody, unknown> = asy
             throw createHttpError(404, "Task not found!");
         }
 
+        if (!task.userId.equals(authenticateUserId)) {
+            throw createHttpError(401, "You cannot access to this task!");
+        }
+
         task.title = updatedTitle;
         task.description = updatedDescription;
 
         const updatedTask = await task.save();
         res.status(200).json(updatedTask);
-
     } catch (error) {
         next(error);
     }
@@ -100,10 +132,18 @@ export const patchTask: RequestHandler<Params, unknown, TaskBody, unknown> = asy
 /**
  * DELETE
  */
-export const deleteTask: RequestHandler<Params, unknown, unknown, unknown> = async (req, res, next) => {
+export const deleteTask: RequestHandler<
+    Params,
+    unknown,
+    unknown,
+    unknown
+> = async (req, res, next) => {
     const taskId = req.params.taskId;
+    const authenticateUserId = req.session.userId;
 
     try {
+        assertIsDefined(authenticateUserId);
+
         if (!mongoose.isValidObjectId(taskId)) {
             throw createHttpError(400, "Invalid task Id");
         }
@@ -112,6 +152,10 @@ export const deleteTask: RequestHandler<Params, unknown, unknown, unknown> = asy
 
         if (!task) {
             throw createHttpError(404, "Task is not found!");
+        }
+
+        if (!task.userId.equals(authenticateUserId)) {
+            throw createHttpError(401, "You cannot access to this task!");
         }
 
         await Task.deleteOne({ _id: taskId });
